Added XML Table of Contents
Document look-up is significantly faster if you grab all the key
elements from a particular document (title, author, id) and store them
in a XML document. Then, when you want to display the table of
contents, you don’t have to query the Exist server. All the information
you need is in one place. The result is a significantly faster backend.